rb-200 branch merge in trunk commits r27003 to r27085 to get XML schema changes into production.
Most of these are to do with either:

(a) a wide ranging documentation update or
(b) packing up the validation recalculation results for the whole PDB (r27044)
    src/python/jiffys/cp_validator_results_to_ftp_dir_structure_and_gz.py

and so have no effect on the actual running of the validation pipeline.

The important (but minor) functional change is from r27004 and r27006
to XML schema. The files changed are:

	src/python/writers/xmlwriter.py
	data/xmlschema/validschema.xsd

result is the output XML file will start with the correct up to date schema location

	<?xml version="1.0" ?>
	<wwPDB-validation-information xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" 
  		xsi:noNamespaceSchemaLocation="http://wwpdb.org/validation/schema/wwpdb_validation_v002.xsd">

rather than the old (mangled) location. The schema is now actually published at the url
given (rather than somewhere else!).