Skip eds and xtriage steps if the structure not "isXray"
This is to deal EM entries with structure factor files trying to run xtriage and eds if default runmany.py without any arguments is run.

1d3e 1d3i 1gr5 1dyl 3epd 3epc 3epf 1m0f codes are an example.

For 3epc get log message:

....
LOG|26-Apr-2016 16:24:56|WARNING    Skipping step xtriage because this is not an X-ray structure
....

LOG|26-Apr-2016 16:26:36|WARNING    Skipping step eds because this is not an X-ray structure

after the change.