Refactoring of chemicalShifts.py and chemicalShiftsConstants.py to be a bit more pep8-like: clean indentation, separating multistatement lines, change s->self.
Despite agreement to not commbine refactoring with changes, also added a change to stop issuing a warning when an atom is present in chemical shifts and in the coordinates, but may not be included in the chemcomp definisitions. Mostly affects positively charged non-stardard residues, where the extra proton is not part of the chem comp definition.

Addresses wwPDB JIRA DAOTHER-1761